backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.18

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Redis & Task Queue
redis>=5.2.0
celery[redis]>=5.4.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Browser Automation
playwright>=1.49.0

# Image Processing
Pillow>=11.0.0

# HTTP Client
httpx>=0.28.0

# Config & Env
pydantic-settings>=2.7.0

# Utilities
python-dateutil>=2.9.0

# Logging
loguru>=0.7.3
